Minor code refactoring, transferred implementations to ZCashClientCaller
