+++ installer/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.zencash</groupId>
         <artifactId>ZENCashOfficialSwingWalletUI</artifactId>
-        <version>0.82.9-SNAPSHOT</version>
+        <version>0.83.0-SNAPSHOT</version>
     </parent>
 
   <artifactId>installer</artifactId>
