+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.zencash</groupId>
     <artifactId>ZENCashOfficialSwingWalletUI</artifactId>
-    <version>0.82.9-SNAPSHOT</version>
+    <version>0.83.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>ZENCash Official Swing Wallet UI</name>
