+++ src/java/com/vaklinov/zcashui/ZCashClientCaller.java
@@ -29,24 +29,11 @@
 package com.vaklinov.zcashui;
 
 
-import java.io.File;
-import java.io.IOException;
-import java.io.LineNumberReader;
-import java.io.StringReader;
+import java.io.*;
 import java.math.BigDecimal;
-import java.util.Date;
-import java.util.HashMap;
+import java.util.*;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Set;
-import java.util.Arrays;
 
 import com.eclipsesource.json.Json;
 import com.eclipsesource.json.JsonArray;
@@ -95,6 +82,9 @@ public class ZCashClientCaller
 		}
 	}
 
+	// ZenCash Advanced settings variables
+	public String commands = "";
+	public boolean runOnce = applyZendCommands();
 
 	// ZCash client program and daemon
 	private File zcashcli, zcashd;
@@ -110,8 +100,6 @@ public class ZCashClientCaller
 	private Map<String, String> transactionConfirmations = Collections.synchronizedMap(
 		new HashMap<String, String>());
 	private long lastTransactionConfirmationsAccess = System.currentTimeMillis();
-
-	
 
 	public ZCashClientCaller(String installDir)
 		throws IOException
@@ -151,10 +139,10 @@ public class ZCashClientCaller
 		throws IOException, InterruptedException 
 	{
 		String exportDir = OSUtil.getUserHomeDirectory().getCanonicalPath();
-		
-        String[] zend_commands = StartupProgressDialog.commands.isEmpty() ?
+
+        String[] zend_commands = commands.isEmpty() ?
 			new String[] {zcashd.getCanonicalPath(), "-exportdir=" + wrapStringParameter(exportDir)} :
-			new String[] {zcashd.getCanonicalPath(), "-exportdir=" + wrapStringParameter(exportDir), StartupProgressDialog.commands};
+			new String[] {zcashd.getCanonicalPath(), "-exportdir=" + wrapStringParameter(exportDir), commands};
 		
 		CommandExecutor starter = new CommandExecutor(zend_commands);
 		Log.info("Starting " + Arrays.toString(zend_commands));
@@ -1332,4 +1320,52 @@ public class ZCashClientCaller
 			map.put(name, val.toString());
 		}
 	}
+
+	public boolean applyZendCommands() {
+		Properties config_file = new Properties();
+		boolean run_option = false;
+		try {
+			config_file.load(new FileInputStream(OSUtil.getSettingsDirectory() + File.separator + "commands.conf"));
+			commands = config_file.getProperty("ZendCommands");
+			run_option = config_file.getProperty("RunOnce").trim().equals("1") ? true : false;
+
+			if (run_option){
+				Log.info("Resetting config file: " + run_option);
+				setConfig();
+			}
+
+		}catch(IOException e){
+			Log.error("Unexpected IO Error:", e);
+		}
+
+		return run_option;
+	}
+
+	public void createConfig(){
+		try {
+			File commandsFile = new File(OSUtil.getSettingsDirectory() + File.separator + "commands.conf");
+			if (!commandsFile.exists()) {
+				setConfig();
+			}
+		}catch (IOException e){
+			Log.error("Unexpected IO Error:", e);
+		}
+
+	}
+
+	public void setConfig() {
+		try {
+			Properties prop = new Properties();
+
+			//Set properties value.
+			prop.setProperty("ZendCommands", "");
+			prop.setProperty("RunOnce", "1");
+
+			prop.store(new FileOutputStream(OSUtil.getSettingsDirectory() + File.separator + "commands.conf"), null);
+
+		} catch (IOException e) {
+			Log.error("Unexpected IO Error:", e);
+		}
+	}
+
 }
