+++ src/resources/messages/zencash_en.properties
@@ -1,4 +1,4 @@
-main.frame.title=ZENCash Desktop GUI Wallet 0.82.9
+main.frame.title=ZENCash Desktop GUI Wallet 0.83.0
 main.frame.progressbar=Starting GUI wallet...
 main.frame.title.testnet=\u0020[using TESTNET]
 main.frame.tab.overview.title=Overview\u0020
