+++ zencash-wallet-swing/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>com.zencash</groupId>
         <artifactId>ZENCashOfficialSwingWalletUI</artifactId>
-        <version>0.82.9-SNAPSHOT</version>
+        <version>0.83.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>zencash-wallet-swing</artifactId>
@@ -178,4 +178,4 @@
             </plugin>
         </plugins>
     </build>
-</project>
+</projec
