+++ zencash-wallet-swing/src/deb/control/control
@@ -1,5 +1,5 @@
 Package: zencash-desktop-gui-wallet
-Version: 0.82.9
+Version: 0.83.0
 Section: misc
 Priority: low
 Architecture: all
